+++ docs/sources/CHANGELOG.md
@@ -18,6 +18,7 @@ The CHANGELOG for the current development version is available at
 
 - A meaningful error message is now raised when a cross-validation generator is used with `SequentialFeatureSelector` ([#377](https://github.com/rasbt/mlxtend/pull/377))
 
+- ColumnSelector now works with Pandas dataframes columns
 
 ##### Changes